Hack MismatchedKeyTypesSqlServerTest for distributed transactions
We have unwanted, intermittent escalation to a distributed transaction
in MismatchedKeyTypesSqlServerTest. We have an event handler which
throws an exception in this case, but a System.Transactions bug makes
TransactionScope.Dispose throw another exception in that case, hiding
the original exception.

This hack removes the call to TransactionScope.Dispose() if an
exception occurs during these tests, allowing the original exception
to bubble up.