Fix to #28648 - Json/Query: translate element access of a json array
Converting indexer into AsQueryable().ElementAt(x) so that nav expansion can understand it and inject MaterializeCollectionNavigationExpression in the right places.
Then in translation we recognize the pattern and convert it back to element access on a JsonQueryExpression.
In order to shape the results correctly, we need to add all the array indexes (that are not constants) to the projection, so that we can populate the ordinal keys correctly (and also to do de-duplication).

Fixes #28648