Fix to #30589 - Query/Json: improve JSON entity/collection comparisons to null (fails currently for optional dependent with all nullable properties)
Handling of optional dependent json entity was incorrect. We were using the same logic as for regular entities, however in case of JSON we don't have the same issues.
We can simply compare the json column to null, or navigate to appropriate json property inside the column and compare that to null.
Additionally, this should give exactly the same results as LINQ to objects - we don't have the problem where we can't distinguish null entity from entity with all values being null, if they are all optional.

Fixes #30589